v2.7.2 Fix jobs to run independent of the project
