+++ src/jobs/tournament_create/workday_tournament_create.py
@@ -2,9 +2,24 @@ import json
 
 import requests
 
-from config.strings import text_automate_webhook_not_work
-from network.api.discord_webhook.http import http_channel_webhook, json_header, content_param
+# Job scripts do not support global imports, must be in the same level as the script file
+from config import *
 
+
+def main():
+    print("Sending signal to Discord Webhook")
+    send_request_create_daily_tournament(get_workday_tournament_content_creation())
+
+def get_workday_tournament_content_creation():
+    command = command_tournament_create
+    tournament_name = weekly_tournament_name
+    team_name = bot_team_name
+    data = f"{command} {tournament_name}, {team_name} Social:\n"
+    if team_whatsapp is not None:
+        data += f"WhatsApp: {team_whatsapp}"
+    if team_discord is not None:
+        data += f"\n\nDiscord: {team_discord}"
+    return data
 
 def send_request_create_daily_tournament(content):
     print("Send request to Discord Channel Webhook to create daily tournaments...")
@@ -31,3 +46,6 @@ def send_request_create_daily_tournament(content):
     except requests.exceptions.RequestException as err:
         print(err)
     return text_automate_webhook_not_work
+
+if __name__ == '__main__':
+    main()
