+++ src/version.py
@@ -1,1 +1,1 @@
-__version__ = '2.7.1'
+__version__ = '2.7.2'
